프론트에서 user id로 받아오던 정보를 authorization 헤더 정보로 전부 교체

+++ src/main/java/kr/megaptera/wherewego/services/DeleteChildService.java
@@ -1,6 +1,8 @@
 package kr.megaptera.wherewego.services;
 
 import kr.megaptera.wherewego.dtos.*;
+import kr.megaptera.wherewego.exceptions.*;
+import kr.megaptera.wherewego.models.*;
 import kr.megaptera.wherewego.repositories.*;
 import org.springframework.stereotype.*;
 import org.springframework.transaction.annotation.*;
@@ -9,14 +11,25 @@ import org.springframework.transaction.annotation.*;
 @Transactional
 public class DeleteChildService {
     private final ChildRepository childRepository;
+    private final UserRepository userRepository;
 
-    public DeleteChildService(ChildRepository childRepository) {
+    public DeleteChildService(ChildRepository childRepository,
+                              UserRepository userRepository) {
         this.childRepository = childRepository;
+        this.userRepository = userRepository;
     }
 
-    public void delete(ChildDeleteDto childDeleteDto) {
+    public void delete(String socialLoginId, ChildDeleteDto childDeleteDto) {
+        User foundUser = userRepository.findBySocialLoginId(socialLoginId)
+            .orElseThrow(UserNotFoundException::new);
+
         Long childId = childDeleteDto.getChildId();
 
-        childRepository.deleteById(childId);
+        Child foundChild = childRepository.findById(childId)
+            .orElseThrow(ChildNotFoundException::new);
+
+        if (foundChild.userId().equals(foundUser.id())) {
+            childRepository.deleteById(childId);
+        }
     }
 }
